# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Current user, their companies, and active project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/active-workspace:
    post:
      operationId: setActiveWorkspace
      tags: [me]
      summary: Set the signed-in user's active (company, project) workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveWorkspaceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          description: Invalid workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/summary:
    get:
      operationId: getMyProjectSummary
      tags: [projects]
      summary: Project summary for the signed-in user's active project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No accessible project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/milestones/upcoming:
    get:
      operationId: getMyUpcomingMilestones
      tags: [milestones]
      summary: Upcoming milestones for the signed-in user's company on the active project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UpcomingMilestone"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/impacts/pending:
    get:
      operationId: getMyPendingImpacts
      tags: [impacts]
      summary: Pending impacts for the signed-in user's company on the active project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingImpact"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{projectId}/milestones:
    get:
      operationId: getProjectMilestones
      tags: [milestones]
      summary: Master schedule – all milestones for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PmMilestone" }
...
  /impacts/{impactId}/respond:
    post:
      operationId: respondToImpact
      tags: [impacts]
      summary: Submit a response to an impact notification
      parameters:
        - in: path
          name: impactId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImpactResponseInput"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneImpact"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    StageCode:
      type: string
      enum:
        - ST1_PRE_DESIGN_CONCEPT
        - ST2_DESIGN_DEVELOPMENT
        - ST3_AUTHORITY_APPROVALS
        - ST4_DETAILED_DESIGN_TENDER
        - ST5_CONSTRUCTION_SHELL_CORE
        - ST6_CONSTRUCTION_MEP_BLOCKWORK
        - ST7_INTERIOR_FITOUT
        - ST8_EXTERNAL_WORKS_FINAL_MEP
        - ST9_COMPLETION_SNAGGING_HANDOVER
        - ST10_CLIENT_HANDOVER_DLP
    MilestoneStatus:
      type: string
      enum: [Planned, OnTrack, AtRisk, Delayed, Completed]
    ImpactRiskLevel:
      type: string
      enum: [None, Low, Medium, High]
    ImpactRiskType:
      type: string
      enum: [Time, Cost, Quality, Scope, Resources, Authority, Other]
    ImpactResponseStatus:
      type: string
      enum: [Pending, Submitted, Reviewed, Closed]
    ChangeEventStatus:
      type: string
      enum:
        - Draft
        - SentForClientReview
        - ClientApproved
        - ClientRejected
        - PMApproved
        - Cancelled
    StageInfo:
      type: object
      required: [stageCode, name, order, milestoneCount, keyOutputCount, companyMilestoneCount]
      properties:
        stageCode: { $ref: "#/components/schemas/StageCode" }
        name: { type: string }
        order: { type: integer }
        milestoneCount: { type: integer }
        keyOutputCount: { type: integer }
        companyMilestoneCount: { type: integer }
    ProjectSummary:
      type: object
      required: [projectId, projectName, projectCode, stages, upcomingMilestoneCount, pendingImpactCount]
      properties:
        projectId: { type: integer }
        projectName: { type: string }
        projectCode: { type: string }
        stages:
          type: array
          items: { $ref: "#/components/schemas/StageInfo" }
        upcomingMilestoneCount: { type: integer }
        pendingImpactCount: { type: integer }
    UpcomingMilestone:
      type: object
      required:
        - id
        - code
        - name
        - stageCode
        - stageName
        - currentDate
        - status
        - isKeyOutput
        - criticalFlag
        - ownerRole
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        stageCode: { $ref: "#/components/schemas/StageCode" }
        stageName: { type: string }
        currentDate: { type: string, format: date-time }
        baselineDate: { type: string, format: date-time }
        previousDate: { type: ["string", "null"], format: date-time }
        status: { $ref: "#/components/schemas/MilestoneStatus" }
        isKeyOutput: { type: boolean }
        criticalFlag: { type: boolean }
        ownerRole: { type: string }
    PendingImpact:
      type: object
      required:
        - id
        - milestoneId
        - milestoneName
        - milestoneCode
        - stageCode
        - stageName
        - oldDate
        - newDate
        - changeReason
        - changeEventStatus
        - notifiedAt
        - responseStatus
      properties:
        id: { type: integer }
        milestoneId: { type: integer }
        milestoneName: { type: string }
        milestoneCode: { type: string }
        stageCode: { $ref: "#/components/schemas/StageCode" }
        stageName: { type: string }
        oldDate: { type: string, format: date-time }
        newDate: { type: string, format: date-time }
        changeReason: { type: string }
        changeEventStatus: { $ref: "#/components/schemas/ChangeEventStatus" }
        notifiedAt: { type: string, format: date-time }
        responseStatus: { $ref: "#/components/schemas/ImpactResponseStatus" }
        isKeyOutput: { type: boolean }
...
    PmMilestoneCompany:
      type: object
      required: [projectCompanyId, companyId, name, role]
      properties:
        projectCompanyId: { type: integer }
        companyId: { type: integer }
        name: { type: string }
        role: { type: string }
    PmMilestone:
      type: object
      required:
        - id
        - code
        - name
        - stageCode
        - stageName
        - stageOrder
        - status
        - ownerRole
        - currentDate
        - baselineDate
        - isKeyOutput
        - criticalFlag
        - isPaymentTrigger
        - owningCompanies
        - contributorCompanies
        - openChangeEventCount
        - pendingResponseCount
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        stageCode: { $ref: "#/components/schemas/StageCode" }
        stageName: { type: string }
        stageOrder: { type: integer }
        status: { $ref: "#/components/schemas/MilestoneStatus" }
        ownerRole: { type: string }
        currentDate: { type: string, format: date-time }
        baselineDate: { type: string, format: date-time }
        previousDate: { type: ["string", "null"], format: date-time }
        isKeyOutput: { type: boolean }
        criticalFlag: { type: boolean }
        isPaymentTrigger: { type: boolean }
        owningCompanies:
          type: array
          items: { $ref: "#/components/schemas/PmMilestoneCompany" }
        contributorCompanies:
          type: array
          items: { $ref: "#/components/schemas/PmMilestoneCompany" }
        openChangeEventCount: { type: integer }
        pendingResponseCount: { type: integer }
...
    ImpactResponseInput:
      type: object
      required: [impactRiskLevel, impactRiskType, mainRiskIssue]
      properties:
        impactRiskLevel: { $ref: "#/components/schemas/ImpactRiskLevel" }
        impactRiskType: { $ref: "#/components/schemas/ImpactRiskType" }
        mainRiskIssue: { type: string, minLength: 1 }
        detailedComment: { type: string }
    MilestoneImpact:
      type: object
      required:
        - id
        - changeEventId
        - milestoneId
        - projectCompanyId
        - oldDate
        - newDate
        - notifiedAt
        - responseStatus
      properties:
        id: { type: integer }
        changeEventId: { type: integer }
        milestoneId: { type: integer }
        projectCompanyId: { type: integer }
        oldDate: { type: string, format: date-time }
        newDate: { type: string, format: date-time }
        notifiedAt: { type: string, format: date-time }
        respondedAt: { type: ["string", "null"], format: date-time }
        impactRiskLevel: { $ref: "#/components/schemas/ImpactRiskLevel" }
        impactRiskType: { $ref: "#/components/schemas/ImpactRiskType" }
        mainRiskIssue: { type: ["string", "null"] }
        detailedComment: { type: ["string", "null"] }
        responseStatus: { $ref: "#/components/schemas/ImpactResponseStatus" }
        pmResolutionNote: { type: ["string", "null"] }
    CompanyMembership:
      type: object
      required: [companyId, companyName, role]
      properties:
        companyId: { type: integer }
        companyName: { type: string }
        role: { type: string }
    Workspace:
      type: object
      required: [companyId, companyName, projectId, projectName, projectCode]
      properties:
        companyId: { type: integer }
        companyName: { type: string }
        projectId: { type: integer }
        projectName: { type: string }
        projectCode: { type: string }
    SetActiveWorkspaceInput:
      type: object
      required: [companyId, projectId]
      properties:
        companyId: { type: integer }
        projectId: { type: integer }
    Me:
      type: object
      required:
        - userId
        - clerkUserId
        - companies
        - workspaces
        - isCompanyAdmin
        - hasAnyAccess
        - pmProjectIds
        - contractorProjectIds
        - adminProjectIds
      properties:
        userId: { type: integer }
        clerkUserId: { type: string }
        email: { type: ["string", "null"] }
        companies:
          type: array
          items: { $ref: "#/components/schemas/CompanyMembership" }
        workspaces:
          type: array
          items: { $ref: "#/components/schemas/Workspace" }
        activeProjectId: { type: ["integer", "null"] }
        activeProjectName: { type: ["string", "null"] }
        activeProjectCode: { type: ["string", "null"] }
        activeCompanyId: { type: ["integer", "null"] }
        activeCompanyName: { type: ["string", "null"] }
        isCompanyAdmin: { type: boolean }
        hasAnyAccess: { type: boolean }
        pmProjectIds:
          type: array
          items: { type: integer }
        contractorProjectIds:
          type: array
          items: { type: integer }
        adminProjectIds:
          type: array
          items: { type: integer }
...